requirements.txt
torch
torchvision
flask
flask-cors
Pillow
numpy
waitress
